feat: expose function from useMollie composable to load the Mollie script async
